backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
sqlalchemy[asyncio]==2.0.35
asyncpg>=0.29.0
alembic==1.13.0
slowapi==0.1.9
bcrypt==4.2.0
pydantic[email]==2.9.0
authlib==1.3.2
httpx==0.27.0
python-jose[cryptography]==3.3.0
python-multipart==0.0.9
itsdangerous==2.2.0
